pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qwen-whitebox"
version = "1.0.0"
description = "White-boxing memory modules of Qwen2.5-0.5B-Instruct"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "533220@qq.com"},
]
keywords = ["llm", "white-box", "memory", "retrieval", "qwen", "faiss"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "faiss-cpu>=1.7.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.66.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest>=7.4.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]
gpu = ["faiss-gpu>=1.7.0"]
cpu = ["faiss-cpu>=1.7.0"]

[project.urls]
Homepage = "https://github.com/533220@qq.com/qwen-whitebox-experiment"
Documentation = "https://github.com/533220@qq.com/qwen-whitebox-experiment#readme"
Issues = "https://github.com/533220@qq.com/qwen-whitebox-experiment/issues"
Source = "https://github.com/533220@qq.com/qwen-whitebox-experiment"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[project.scripts]
qwen-whitebox-train = "scripts.train_adapter:main"
qwen-whitebox-build-index = "scripts.build_faiss:main"
qwen-whitebox-demo = "scripts.editable_demo:main"